feat: Rename Social Plugin packages - MEED-8936 - Meeds-io/MIPs#192
