+++ pwa-plugin/src/main/java/io/meeds/pwa/plugin/social/SpaceInvitationReceivedPwaNotificationPlugin.java
@@ -32,7 +32,7 @@ import io.meeds.portal.permlink.model.PermanentLinkObject;
 import io.meeds.portal.permlink.service.PermanentLinkService;
 import io.meeds.pwa.model.PwaNotificationMessage;
 import io.meeds.pwa.plugin.PwaNotificationPlugin;
-import io.meeds.social.permlink.plugin.SpacePermanentLinkPlugin;
+import io.meeds.social.space.plugin.SpacePermanentLinkPlugin;
 
 import lombok.Getter;
